wpt_babel/content-security-policy/unsafe-hashes/javascript_src_denied_wrong_hash-window_open.html

<!DOCTYPE HTML>
<html>
 <head>
  <meta content="script-src 'unsafe-hashes' 'nonce-abc'
    'sha256-VjH6k67F4kobUnNDOBE85QiJ9cuZMiYT6desKXvezVg=';" http-equiv="Content-Security-Policy"/>
  <!--
    'sha256-IIiAJ8UuliU8o1qAv6CV4P3R8DeTf/v3MrsCwXW171Y=' ==> 'javascript:opener.postMessage('pass', '*')'
    -->
  <script nonce="abc" src="/resources/testharness.js">
  </script>
  <script nonce="abc" src="/resources/testharnessreport.js">
  </script>
 </head>
 <body>
  <div id="log">
  </div>
  <script nonce="abc">
   "use strict";

var t1 = async_test("Test that the javascript: src is not allowed to run");
window.onmessage = t1.unreached_func("Should have not received any message");
window.addEventListener('securitypolicyviolation', t1.step_func_done(function (e) {
  assert_equals(e.violatedDirective, 'script-src-elem');
  assert_equals(e.blockedURI, 'inline');
}));
window.open("javascript:opener.postMessage('pass', '*')");
  </script>
 </body>
</html>
